GROOVY-4694: Move AstBuilderTransformation Global xForm to separate module
