+++ settings.gradle
@@ -30,6 +30,7 @@ gradleEnterprise {
 }
 
 def subprojects = ['groovy-ant',
+        'groovy-astbuilder',
         'groovy-bsf',
         'groovy-cli-commons',
         'groovy-cli-picocli',
@@ -43,6 +44,7 @@ def subprojects = ['groovy-ant',
         'groovy-jmx',
         'groovy-json',
         'groovy-jsr223',
+        'groovy-macro',
         'groovy-nio',
         'groovy-servlet',
         'groovy-sql',
@@ -52,7 +54,6 @@ def subprojects = ['groovy-ant',
         'groovy-test-junit5',
         'groovy-testng',
         'groovy-xml',
-        'groovy-macro',
         'groovy-yaml',
         'performance',
         'binary-compatibility'
