+++ src/spec/test/CustomizersTest.groovy
@@ -102,25 +102,6 @@ class CustomizersTest extends GroovyTestCase {
         '''
     }
 
-    void testAstTransformationCustomizerWithClosureExpression() {
-        // tag::ast_cz_closure[]
-        def configuration = new CompilerConfiguration()
-        def expression = new AstBuilder().buildFromCode(CompilePhase.CONVERSION) { -> true }.expression[0]
-        def customizer = new ASTTransformationCustomizer(ConditionalInterrupt, value: expression, thrown: SecurityException)
-        configuration.addCompilationCustomizers(customizer)
-        def shell = new GroovyShell(configuration)
-        shouldFail(SecurityException) {
-            shell.evaluate("""
-                // equivalent to adding @ConditionalInterrupt(value={true}, thrown: SecurityException)
-                class MyClass {
-                    void doIt() { }
-                }
-                new MyClass().doIt()
-            """)
-        }
-        // end::ast_cz_closure[]
-    }
-
     void testSecureASTCustomizer() {
         // tag::secure_cz[]
         def scz = new SecureASTCustomizer()
