+++ src/test/org/codehaus/groovy/control/customizers/ASTTransformationCustomizerTest.groovy
@@ -20,29 +20,20 @@ package org.codehaus.groovy.control.customizers
 
 import groovy.test.GroovyTestCase
 import groovy.transform.TimedInterrupt
+import groovy.util.logging.Log
+import org.codehaus.groovy.ast.ASTNode
 import org.codehaus.groovy.ast.ClassHelper
 import org.codehaus.groovy.ast.expr.ClassExpression
 import org.codehaus.groovy.ast.expr.PropertyExpression
+import org.codehaus.groovy.control.CompilePhase
 import org.codehaus.groovy.control.CompilerConfiguration
-import groovy.util.logging.Log
-
-import java.util.concurrent.TimeUnit
-import java.util.logging.Logger
+import org.codehaus.groovy.control.SourceUnit
 import org.codehaus.groovy.transform.ASTTransformation
 import org.codehaus.groovy.transform.GroovyASTTransformation
-import org.codehaus.groovy.control.CompilePhase
+
+import java.util.concurrent.TimeUnit
 import java.util.concurrent.atomic.AtomicBoolean
-import org.codehaus.groovy.ast.ASTNode
-import org.codehaus.groovy.control.SourceUnit
-import java.lang.annotation.Retention
-import java.lang.annotation.Target
-import org.codehaus.groovy.transform.GroovyASTTransformationClass
-import java.lang.annotation.ElementType
-import java.lang.annotation.RetentionPolicy
-import org.codehaus.groovy.ast.ClassNode
-import org.objectweb.asm.Opcodes
-import org.codehaus.groovy.ast.builder.AstBuilder
-import groovy.transform.ConditionalInterrupt
+import java.util.logging.Logger
 
 /**
  * Tests the {@link ASTTransformationCustomizer}.
@@ -94,62 +85,6 @@ class ASTTransformationCustomizerTest extends GroovyTestCase {
         customizer = new ASTTransformationCustomizer(Log)
         shouldFail(IllegalArgumentException) {
             customizer.annotationParameters = [invalid: 'logger']
-        }
-    }
-
-    void testLocalTransformationWithClosureAnnotationParameter() {
-        // add @Contract({distance = 1 })
-        customizer = new ASTTransformationCustomizer(Contract)
-        final expression = new AstBuilder().buildFromCode(CompilePhase.CONVERSION) {->
-            distance = 1
-        }.expression[0]
-        customizer.annotationParameters = [value: expression]
-        configuration.addCompilationCustomizers(customizer)
-        def shell = new GroovyShell(configuration)
-        def result = shell.evaluate("""
-            class MyClass {
-                int distance
-                MyClass() {}
-            }
-            new MyClass()
-        """)
-        assert result.distance == 1
-    }
-
-    void testLocalTransformationWithClosureAnnotationParameter_notAnnotatedAsASTInterface() {
-        // add @Contract2({distance = 1 })
-        customizer = new ASTTransformationCustomizer(Contract2, "org.codehaus.groovy.control.customizers.ContractAnnotation")
-        final expression = new AstBuilder().buildFromCode(CompilePhase.CONVERSION) {->
-            distance = 1
-        }.expression[0]
-        customizer.annotationParameters = [value: expression]
-        configuration.addCompilationCustomizers(customizer)
-        def shell = new GroovyShell(configuration)
-        def result = shell.evaluate("""
-            class MyClass {
-                int distance
-                MyClass() {}
-            }
-            new MyClass()
-        """)
-        assert result.distance == 1
-    }
-
-    void testLocalTransformationWithClassAnnotationParameter() {
-        // add @ConditionalInterrupt(value={ true }, thrown=SecurityException)
-        final expression = new AstBuilder().buildFromCode(CompilePhase.CONVERSION) {->
-            true
-        }.expression[0]
-        customizer = new ASTTransformationCustomizer(ConditionalInterrupt, value:expression, thrown:SecurityException)
-        configuration.addCompilationCustomizers(customizer)
-        def shell = new GroovyShell(configuration)
-        shouldFail(SecurityException) {
-            shell.evaluate("""
-                class MyClass {
-                    void doIt() { }
-                }
-                new MyClass().doIt()
-            """)
         }
     }
 
@@ -217,26 +152,4 @@ interrupted'''
         
     }
 
-}
-
-@Retention(RetentionPolicy.SOURCE)
-@Target([ElementType.TYPE])
-@GroovyASTTransformationClass("org.codehaus.groovy.control.customizers.ContractAnnotation")
-protected @interface Contract {
-    Class value();
-}
-
-@GroovyASTTransformation(phase=CompilePhase.CONVERSION)
-protected class ContractAnnotation implements ASTTransformation, Opcodes {
-    void visit(ASTNode[] nodes, SourceUnit source) {
-        def node = nodes[0]
-        def member = node.getMember("value")
-        ((ClassNode)nodes[1]).getDeclaredConstructors()[0].code = member.code
-    }
-}
-
-@Retention(RetentionPolicy.SOURCE)
-@Target([ElementType.TYPE])
-protected @interface Contract2 {
-    Class value();
 }
