+++ subprojects/groovy-astbuilder/src/main/resources/META-INF/services/org.codehaus.groovy.transform.ASTTransformation
@@ -13,5 +13,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# global transformation to handle @Grab annotation
-groovy.grape.GrabAnnotationTransformation
+#global transformation for AST Builder
+org.apache.groovy.ast.builder.AstBuilderTransformation
