+++ subprojects/groovy-astbuilder/src/test/groovy/bugs/Groovy4272Bug.groovy
