+++ subprojects/groovy-astbuilder/src/test/groovy/bugs/MyConstantsASTTransformation4272.groovy
